Created seprate files to handle displaying videos on the gallary
